+++ src/pygeoprocessing/__init__.py
@@ -11,6 +11,7 @@ import pkg_resources.extern.packaging.version
 from . import geoprocessing
 from .geoprocessing import _assert_is_valid_pixel_size
 from .geoprocessing import align_and_resize_raster_stack
+from .geoprocessing import build_overviews
 from .geoprocessing import calculate_disjoint_polygon_set
 from .geoprocessing import convolve_2d
 from .geoprocessing import create_raster_from_bounding_box