+++ HISTORY.rst
@@ -3,6 +3,11 @@ Release History
 
 Unreleased Changes
 ------------------
+* Added a function to build overviews for a raster,
+  ``pygeoprocessing.build_overviews``. Related to this,
+  ``pygeoprocessing.get_raster_info()`` now includes an ``'overviews'`` key
+  listing the pixel dimensions of each overview layer in a raster.
+  https://github.com/natcap/pygeoprocessing/issues/280
 * Added a D8 stream extraction function at
   ``pygeoprocessing.routing.extract_streams_d8`` which takes a D8 flow
   accumulation raster and a flow accumulation threshold, setting all pixels